requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
requests
pytest
httpx
pytest-playwright
playwright
alembic